Moved to a 'cloning' approach to creating new N_Vectors within the SUNDIALS solvers.
Removed NVSpec structure. The list of operations is now in the N_Vector structure.

The implementation specific NVSpec_Init function is now replaced by an
implementation-specific N_VNew_Impl function which creates a new N_Vector.
This vector is used as needed as a template for cloning new ones.

N_VGetData and N_VSetData were replaced with N_VGetArrayPointer and
N_VSetArrayPointer, respectively. The functionality is basically the same as
before, but I moved back to explicitely pass pointers to realtype (and not
pointers to void). The new names are hopefully more illuminating as to
what these functions are supposed to do.

Removed N_VMake / N_VDispose from the table of operations. A particular
NVECTOR implementation can (and is encouraged) to provide these functions
but they are not needed by the SUNDIALS solvers anymore (see the interfaces
in CVODE, CVODES, IDA, and IDAS to the dense linear solver).

Changes to comments.