New NVECTOR module: NVECTOR_SPCPARALLEL
As for NVECTOR_PARALLEL, the underlying data is still a contiguous array of
realtype values, but it is structured so that it can be used for 3D simulations
in which several species (nspc) coexist at each spatial grid point. The data
may include ghost boundary grid points and all vector operations are written
so that they operate only on the actual computational grid points (i.e., they
ignore the data at points in the ghost boundary).

The indexes vary in the following order (from fastest to slowest): x location,
y location, z location, species index.
Note that this implies that two species at the same spatial location are not
next to each other in the data array (this arrangement is suitable for CFD-type
calculations, where the species can be velocity components, temperature, pressure,
density, etc.)

Accessor macros and implementation-specific functions provide a transparent
mechanism of dealing with such structures.

All F2C interface functions are provided.

Notes:
  1. The ghost boundary widths can be zero in any dimension.
  2. This NVECTOR module should also work for 2D simulations, setting Nz=1
     (width of the computational domain in the z-direction) and NGz=0
     (width of the ghost boundary in the z-direction).