Changes to the build system:
Modify acinclude so that the Fortran name-mangling scheme tests are done
whether FCMIX is enabled or not. This is useful, for example, for the case
of a platform that does not have a Fortran compiler but on which a
precompiled Lapack library exists. The user should be able to specify
the Fortran name-mangling scheme (which he can infer by looking at
liblapack with either nm or objdump).

Modify ltmain.sh so that it uses the default CC tag if it is unable
to infer one (and disable the resulting error). This issue only arises
in more recent versions of ltmain.sh but the alternative of using
older versions (such as that shipped with sundials-2.3.0) is not
a solution since it introduces a whole lot more problems related
to shared libraries (wrong/no extension, problems recongnizing pgi
compilers, etc)

For each solver, create separate libraries with the Lapack interfaces
(so that we do not have to link the Blas/Lapack libraries to an example
that does not use it. Indeed, if the interfaces are in the main library
and we use shared libraries, we get undefined symbols unless we link Lapack)

Mark all exported functions in the header files (using a preprocessor symbol
defined in sundials_config.h). SUNDIALS_EXPORT is usually empty, except
when building/using shared dll libraries on Windows.

Fix all solvers/examples so that g++ does not choke on functions that take
a void* argument for a function argument

Move sundials_config.in and sundials_config.h into include/sundials, from
of src/sundials

Create a bin directory which contains the templates sundials-config.in
and fortran-update.in. The resulting scripts will be installed in bindir.

Add Makefile_ex.in in all example directories. Modify autotools such that,
if a set of examples is compilable, a sample Makefile_ex is also generated
from Makefile_ex.in. Moreover, modify the corresponding Makefile.in such that,
on make install, we install the sources, sample outputs, and the Makefile_ex
(renamed Makefile). This way, the user is provided with a complete set of
files on which he/she can test the Sundials installation. Moreover, this
is a fix for the current "istallation" of examples which is broken in case
shared libraries are used (indeed, since we link against stuff in the source
tree, after a make clean in the source tree, the installed examples will
cease to work!)