Bug fix:
In the rootfinding for all solvers (CVODE, CVODES, IDA, IDAS, CPODES), fixed
bug (irfnd was not reset to 0 in the solver re-initialization functions) which
resulted in additional unnecessary g function evaluation (in the **Rcheck2 fct)
on a second run after re-initializing the solver right after a return at a root.


New feature:

In CVODE added optional input function CVodeSetRootDirection to specify
the zero-crossing direction. If rootdir[i] > 0, then only roots of g_i
where g_i is increasing are reported. If rootdir[i] < 0, then only roots
of g_i where g_i is decreasing are reported. If rootdir[i] = 0, then all
roots of g_i are reported. The default is rootdir[i]=0, for all i.

In addition, the return int array of CVodeGetRootInfo now also encodes the
zero-crossing direction. If iroots[i]=-1, then g_i had a root and g_i was
decreasing at the root. If iroots[i]=+1, then g_i had a root and g_i was
increasing at the root. If iroots[i]=0, then either g_i did not have a root
or it did, but g_i was varying in the direction opposite to that indicated
by rootdir[i].

This additional feature is useful, for example, for collision problems, where
the g functions are (signed) distances between objects, and only the roots in
the decreasing diection are needed.

After a bit more testing, I'll add this fature to the other solvers as well.