Added optional input to specify directino of zero-crossing direction. If rootdir[i] > 0, then only roots of g_i where g_i is increasing are reported. If rootdir[i] < 0, then only roots of g_i where g_i is decreasing are reported. If rootdir[i] = 0, then all roots of g_i are reported. The default is rootdir[i]=0, for all i.
In addition, the return int array "iroots" now also encodes the
zero-crossing direction. If iroots[i]=-1, then g_i had a root and g_i was
decreasing at the root. If iroots[i]=+1, then g_i had a root and g_i was
increasing at the root. If iroots[i]=0, then either g_i did not have a root
or it did, but g_i was varying in the direction opposite to that indicated
by rootdir[i].