Extended the adjoint sensitivity module to allow FSA on the forward phase - needed for computing Hessian information using FSA+ASA. Both the RHS and quadrature RHS of a backward problem can also depend on forward sensitivities (besides depending on the forward solution). If that is the case, the backward problem or its associated quadrature problem must be initialized using CVodeMallocBs and CVodeQuadMallocBs, respectively. For backward problems that do not depend on forward sensitivities, the previous initialization functions should be used (CVodeMallocB and CVodeQuadMallocB).
To minimize the number of functions in the API, I removed the function
from the argument list of CVodeReInitB and CVodeQuadReInitB so that they
can be used to reinitialize either type of backward problem.

Note that, if FSA is enabled for the forward problem, the sensitivity
vectors are always stored at check points so that the forward integration
could be duplicated on a possible second pass. However, storage and
interpolation of sensitivity variables is controlled by the user: storage
of sensitivity variables in the interpolation structures can be disabled
by the user; interpolation of the sensitivity variables is performed
only if there exists at least one backward problem that requires them.