(1) In case the user provides a function to set the error weights, this function will receive the same user data pointer as the one (optionally) specified through ***SetUserData.
(2) Now, all user-supplied functions receive the same user data
pointer. The only exception is a user-provided error handler
function for which a possibly different data pointer can be
specified (through the ***SetErrHandler function). The reasons
for this are: (1) if a user error handler functin is supplied,
this is usually done as the first call after the ***Create
call, at which point the user_data pointer may not be set yet;
(2) the data needed in the error handler is typically disjunct
from problem data needed elsewhere.

(3) Manyn minor clean-up changes throughout...