SP: Updating documentation build system for KINSOL in 'slaven' branch.
The new documentation builds with xelatex and pdflatex. All references to 
pstricks and epsfig are removed from tex files. Command to print "C" as in 
C-language was changed from \C to \CC. The former symbol conflicts with 
definitions in xelatex and causes havoc when building documentation. Package 
supertabular is replaced by newer xtabular package. A bug that prevented 
crossreferencing to external documents for installation and linear solver have 
been fixed. Images in pdf are added to the repository.