SP: PETSc vector can be created and destroyed from N_Vecor wrapper.
For now PETSc implementation of N_Vector does duplicate operations on Vec 
and on raw data array (as in MPI implementation). This will be kept for code 
verification purposes. As soon as basic vector operations are verified, it will be 
removed. 

A decision needs to be made where to call PetscInitialize() function. Possible 
options are SUNDIALS initialization functions such as IDAInit (or perhaps even 
IDACreate) or the first vector to be instantiated.