SP: Fixed 'clone' function and added 'make' function to PETSc vector wrapper.
1. N_VMake_petsc takes PETSc vector as the argument and creates N_Vector wrapper 
for it. The wrapper does not own the PETSc vector. Upon deletion of the wrapper, 
it is user's responsibility to delete PETSc vector.

2. N_VClone_petsc is now using PETSc function VecDuplicate to clone the wrapped 
PETSc vector. Unlike built in N_Vector types, PETSc vector is not uniquely defined 
by communicator, global and local length. To make a proper clone, one has to call 
VecDuplicate.