SP: Removed unused pointer to raw data in parhyp vector.
Hypre vector wrapper was allocating data block, which was not used and 
separate from data in the Hypre vector. This has been removed in this 
commit.

Macro that returns pointer to raw data of wrapped hypre vector was 
modified to first check if hypre vector is allocated at all to prevent 
irrecoverable segfault error.

Also in this commit, comments in the code are expanded and corrected 
where needed.