SP: Removed N_VNew_ParHyp method.
When this method is used, it is N_Vector wrapper who sets and 
actually owns the partitioning. Having this method in the code 
may leave a wrong impression that SUNDIALS restricts how Hypre 
could be used. Typically one would want to use Hypre partitioning 
utilities, and that becomes quite difficult if the vector is created 
with N_VNew_ParHyp.

It is recommended that user creates partitioning and Hypre vector 
first, and then N_Vector wrapper by invoking N_VMake_ParHyp.