SP: Modified a couple of parhyp examples to be more "hypre like".
In an attempt to make Hypre related examples more "educational" I modified 
a couple of them to use Hypre interface for creating and initializing vectors. 
This is still work in progress. The objective is to use Hypre interface in all user 
supply functions and completely avoid accessing raw data underneath Hypre 
vectors.