SP: Fixing a nasty bug in cmake to find hypre library
On newer linux systems, when linking to static libraries, one needs to add 
explicitly "-lm" to link to standard math library. This is taken care of in the 
main cmake configuration file, but not in cmake scripts that generate 
make files to create library tests that verify found libraries are 
functional. As a result, some static build configurations fail to verify 
and link to libraries such as hypre or KLU. I fixed this for hypre, similar fixes 
are needed for other libraries. Some cleanup of cmake system would 
be helpful.