Fix Bug in Setting MPI_RUN_COMMAND
Update CMake to use the value of MPI_RUN_COMMAND. Before if this
option was not set to srun, then the value of MPI_RUN_COMMAND is
ignored and mpirun is used in the test runner. This causes make
test to fail when the full path to the MPI run command is needed
(e.g. mpirun is not is the users path) or when the command is
not srun or mpirun (e.g. mpiexec with the mpich). Remove warning
message if MPI_RUN_COMMAND is not srun or mpirun.