SUNDIALS-153 Add CMake check for hypre precision
Added check to only enable hypre with double precision. This should
be removed for the next release of hypre when support for single and
extended precision is added.

Added notes that CMake should check that hypre and PETSc were built
with same precision and index type as SUNDIALS is configured with.