Fix bug when findng LAPACK lib with xSDK option in ccmake gui
When automatically finding LAPACK and using xSDK options, we need
to copy the found library path into the xSDK variable so it appears
in the ccmake gui.

This may not be necessary for other packages (except BLAS) because
xSDK has users specify the full path and does not require
automatically finding the library (except BLAS and LAPACK).