Explicitly set include paths for external libraries in makefile/cmake templates
Installed examples would not build with the automatically generated makefile
and cmake files. Examples using external libraries could not find header files
for the external libraries. Template files now explicitly set the include path
for external libraries. This was not a problem previously and these files have
not changed but I do not see how the paths to the header files were found before.