Add SUNDIALS_DEVTESTS flag; Add keyword EXAMPLE_TYPE to SUNDIALS_ADD_TEST
Add check to see if README_DEV exists to determine if this is a released
or development version of Sundials and set the SUNDIALS_DEVTESTS flag
accordingly. If SUNDIALS_DEVTESTS is true, then all examples are
added to the regression tests (i.e. current make test behavior).

Add the keyword argument EXAMPLE_TYPE to the SUNDIALS_ADD_TEST macro to
indicate which examples should NOT be in released regression testing.
If SUNDIALS_DEVTEST if false, then only examples with an empty example
type are added to regression tests. Excluded examples are given the
'develop' example type.

Clean up comments in testRunner script.