Fixes for pointer conversion compiler warnings/errors
In idaa.c we passed ida_mem (of void* type) in as an argument where we
should have passed IDA_mem (of IDAMem type).  This has been in
SUNDIALS for quite some time; I do not know why we are first hearing
of this now.

In the remaining files we used "realloc" to re-allocate memory, but
never cast the resulting (void*) pointer to the appropriate data
type.

Additionally, in implementing this patch I noticed a bug in my
implementation of the sunmatrix_sparse.c file -- if sunindextype is
64-bit, and if we needed to copy a more filled-in matrix into a less
filled-in matrix, the new memory for the indexvals pointer would have
been too short.  This corner case was too esoteric for our testing
harness, but it is now fixed in this patch.