update how name-mangling scheme is set, enable Fortran in main file
* Check for user defined name-mangling scheme first
* Call enable_language(Fortran) in main CMake file if needed
* If user did not define a name-mangling scheme let CMake figure
  it out and create sundials_fcmix.h with the needed macros. These
  macros are currently defined in sundials_config.h

TO DO: Need to udpate the macro names thoughout with _GLOBAL prefix
since cmake creates global and module name-mangling schemes.

NOTE: Regression tests will pass because the #ifdef statements
checking if the name-mangling macro exists will default to lower
case and one underscore if the macro is not defined which works in
our tests.