Remove CMakeDetermine*Compiler from Sundials<lang>.cmake
Using include(CMakeDetermine*Compiler) makes it so enable_language
does not cause a fatal error if the compiler for the language is
broken. This is why we test the CXX and Fortran compilers ourselves
rather than letting CMake do it. This allows us to disable features
if a compiler is not working rather than halting the configuration
which is what CMake would normally do.

Benefits of removing: include(CMakeDetermine*Compiler):
* CMake will test compilers for us
* CMake will check if the Fortran compiler supports F90
* CMake can automatically determine and test Fortran name-mangling
* CMake can create a header file with the name-mangling macros
* CMake FindMPI will find the MPI wrappers for any enabled language

Drawbacks of removing: include(CMakeDetermine*Compiler):
* CMake will stop the configuration if the CXX or Fortran compiler
  does not work (assuming a CXX or Fortran compiler is needed).

Users that will be most impacted by this are those that use LAPACK
and do not have a functioning Fortran compiler. Previosuly, if
LAPACK was enabled and the Fortran compiler could not be found or
did not work the name-mangling scheme would default to lower case
with one underscore (which the user could change). So without a
Fortran compiler and LAPACK enabled the configuration would carry
on without halting.

Now if the user does not set the name-mangling scheme and the
Fortran compiler does not work the configuration will halt rather
going to a default name-mangling scheme. In this case the user
should set a scheme and the configuration will go on without
stopping (assuming Fortran examples are not enabled).

Additionally, FCMIX could be used without a Fortran compiler since
we only need the name-mangling scheme but we disable FCMIX if a
working Fortran compiler is not found. Really we only need a
Fortran compiler to build the Fortran examples.

This change in behavior needs to be discussed before before
incorporating this branch.