Clean up return codes
Functions provided to the generic nonlinear solver (i.e., NLS Sys,
LSetup, and LSolve) now return IDA specific constants which are
passed back to IDA from the generic NLS. The NLS only needs to
know that these are non-zero values. Handle the possibility of a
memory failure in the functions provided to the generic NLS. Use
SUN_NLS_CONV_RECVR for failures due to possible divergence or
reaching the max allowed number of iterations.