remove check that lsolve is non-null in IDA
With NLS modules it is possible for lsolve to be NULL if the NLS
module provides its own linear solver. As such integrators should
not check that lsolve is non-null. NLS modules that require a
sundials LS module need to check that lsolve is non-null and
return an error if appropriate.