Add GetCurIter to the NLS API
Added a function to get the nonlinear iteration index of the current
solve from the nonlinear solver. This is needed to accommodate the
case where a spils solver returns SUNLS_RES_REDUCED on the first
nonlinear solver iteration and the integrator returns success
(CVODE, CVODES, and ARKODE) rather than a recoverable error (IDA
and IDAS).