fix bug in final update and add deubgging options/output
In the serial, cuda, and cuda_serial versions of the example the same
vector is used to store the numerator for the G calculation and the
new value of mu. When the G values are computed the numerator values are
overwritten so later updates use an incorrect numerator value.

Added debugging options to CMake and debugging output to KINFP.