Bugfix: check tstop first (#322)
Fixed a bug in ARKODE, CVODE, CVODES, IDA, and IDAS where the stop time
may not be cleared when using normal mode if the requested output time
is the same as the stop time. Additionally, with ARKODE, CVODE, and
CVODES an unnecessary call to
interpolate the solution at the stop time may occur in this case.