Bugfix: stop time and output time reached in same step (#349)
Fixes #339: A regression was introduced by the stop time bug fix in
v6.6.1 causing integrators to return at the stop time rather than the
requested output time if the stop time was reached in the same step in
which the output time was passed.

Also fixes a missing check in ARKODE to interpolate (or not) the
solution at the stop time.