Bugfix: CMake FindMAGMA (#356)
Always append hipblas and hipsparse CMake targets with HIP backend as these
libraries are not reliably included in the pkgconfig file.

Replace cublas and cusparse with the CMake targets because the library path in
the pkgconfig file is not reliable. Specifically, the path is wrong on systems
using the NVIDIA HPC SDK.

---------

Co-authored-by: David Gardner <gardner48@llnl.gov>