FindKLU.cmake: Avoid detecting MSVC libraries when targeting MinGW
For MinGW, setting the library prefix is not necessary and setting the
library suffix to d.lib is incorrect. The static library suffix is .a for
MinGW.

Prepending those values to the list of library prefixes and suffixes
might not actually cause an issue if no libraries are found that match
those patterns. But slight ABI differences between MSVC and MinGW could
lead to unexpected behavior when mixing binaries of both targets.

Signed-off-by: Markus Mützel <markus.muetzel@gmx.de>