FindKLU.cmake: Avoid detecting MSVC libraries when targeting MinGW (#476)
For MinGW, setting the library prefix is not necessary and setting the
library suffix to d.lib is incorrect. The static library suffix is .a
for MinGW.

Prepending those values to the list of library prefixes and suffixes
might not actually cause an issue if no libraries are found that match
those patterns. But slight ABI differences between MSVC and MinGW could
lead to unexpected behavior when mixing binaries of both targets.

This was already discussed in
https://github.com/LLNL/sundials/pull/407#discussion_r1481939567. But
apparently that change got lost in one of the rebases in that PR.

---------

Signed-off-by: Markus Mützel <markus.muetzel@gmx.de>