Do not cast pointers to arrays of differently sized integers. (#485)
Casting between `int64_t *` and `int32_t *` does *not* maintain the
values in the array. Instead, it tells the compiler to interpret the
memory at that pointer as an array of a different type (i.e., two
`int32_t` elements "become" one `int64_t` element). That can lead to all
kinds of errors and is likely not what was intended.

Remove the pointer casts to allow the compiler to emit an error on
compile-time instead of potentially causing, e.g., an array overflow on
runtime if `sunindextype` has a different size from `KLU_INDEXTYPE`.

---------

Signed-off-by: Markus Mützel <markus.muetzel@gmx.de>
Co-authored-by: Cody Balos <balos1@llnl.gov>