Bugfix: Rootfinding with Fixed Step Size (#614)
Fix a bug when enabling rootfinding with fixed step sizes and the
initial value of the rootfinding function is zero. In this case,
uninitialized right-hand side data (`fn`) is used to compute a forward
Euler step to get a state near the initial condition to determine if any
rootfinding functions are active at the initial condition.